Edit mypage merge

+++ db/migrate/20151008121401_create_items.rb
@@ -2,9 +2,9 @@ class CreateItems < ActiveRecord::Migration
   def change
     create_table :items do |t|
       t.string :name
-      t.integer :spent_men
-      t.integer :spent_women
-      t.integer :spent_child
+      t.float :spent_men
+      t.float :spent_women
+      t.float :spent_child
 
       t.timestamps null: false
     end
